+++ src/SvgHistoryTreeProvider.ts
@@ -1,3 +1,4 @@
+// svg-viewer\src\SvgHistoryTreeProvider.ts
 import * as vscode from 'vscode';
 import * as path from 'path';